+++ src/components/_materialUi/Tooltip/index.stories.template.tsx
@@ -11,7 +11,7 @@ import * as React from 'react';
 import { Story } from '@storybook/react/types-6-0';
 import { Tooltip, TooltipProps } from '@material-ui/core';
 
-import { HelpCircleOutline } from '~/icons';
+import HelpCircleOutline from '~/icons/HelpCircleOutline';
 import { Playground } from '~/helpers/playground';
 
 export const Template: Story<TooltipProps> = args => (