+++ src/styles/themeActEt/overrides.ts
@@ -7,10 +7,12 @@
  * @prettier
  */
 
+import Color from 'color';
 import { Overrides } from '@material-ui/core/styles/overrides';
 
 import COLORS from './colors';
 import DIMS from './dims';
+import SHAPE from './shape';
 
 const OVERRIDES: Overrides = {
   MuiAlert: {
@@ -25,6 +27,11 @@ const OVERRIDES: Overrides = {
     },
     outlinedWarning: {
       color: COLORS.WARNING_DARK,
+    },
+  },
+  MuiBadge: {
+    badge: {
+      fontSize: 10,
     },
   },
   MuiButton: {
@@ -71,6 +78,151 @@ const OVERRIDES: Overrides = {
       '&:active': {
         color: COLORS.PRIMARY_DARK,
       },
+    },
+  },
+  MuiDataGrid: {
+    root: {
+      '& .MuiDataGrid-footerContainer': {
+        '& .MuiTablePagination-root': {
+          '& .MuiToolbar-root': {
+            '& .MuiSelect-root': {
+              fontSize: 14,
+            },
+          },
+        },
+      },
+      '& .MuiDataGrid-main': {
+        '& .MuiDataGrid-columnHeader': {
+          '& .MuiDataGrid-columnHeaderTitleContainer': {
+            '& .MuiDataGrid-columnHeaderTitle': {
+              color: COLORS.BLACK,
+              userSelect: 'none',
+            },
+          },
+          '& .MuiDataGrid-iconButtonContainer': {
+            '& .MuiIconButton-root': {
+              color: Color(COLORS.BLACK).fade(0).rgb().string(),
+              padding: '0 0 1px 0',
+            },
+            '& .MuiIconButton-root[aria-label="Show filters"]': {
+              color: `${COLORS.BLACK} !important`,
+            },
+            alignItems: 'center',
+            display: 'flex',
+          },
+          '& .MuiDataGrid-menuIcon': {
+            '& .MuiIconButton-root[aria-label="Menu"]': {
+              color: Color(COLORS.BLACK).fade(0.35).rgb().string(),
+            },
+          },
+          '&.MuiDataGrid-withBorder': {
+            borderRightColor: COLORS.GREY_300,
+          },
+          outline: 'none',
+        },
+        '& .MuiDataGrid-columnHeader--sorted': {
+          '& .MuiIconButton-root[aria-label="Sort"]': {
+            color: `${COLORS.BLACK} !important`,
+          },
+          backgroundColor: COLORS.GREY_300,
+        },
+        '& .MuiDataGrid-columnHeaderWrapper': {
+          backgroundColor: COLORS.GREY_200,
+        },
+        '& .MuiDataGrid-columnSeparator': {
+          display: 'none',
+        },
+        '& .MuiDataGrid-overlay': {
+          backgroundColor: Color(COLORS.WHITE).fade(0.35).rgb().string(),
+          zIndex: 1,
+        },
+      },
+      '& .MuiDataGrid-toolbarContainer': {
+        '& .MuiButton-root': {
+          marginRight: DIMS.SPACING,
+        },
+        padding: DIMS.SPACING,
+      },
+      '&.primary': {
+        '& .MuiDataGrid-main': {
+          '& .MuiDataGrid-columnHeader': {
+            '& .MuiDataGrid-columnHeaderTitleContainer': {
+              '& .MuiCheckbox-root': {
+                color: COLORS.WHITE,
+              },
+              '& .MuiDataGrid-columnHeaderTitle': {
+                color: COLORS.WHITE,
+              },
+            },
+            '& .MuiDataGrid-iconButtonContainer': {
+              '& .MuiIconButton-root': {
+                color: Color(COLORS.WHITE).fade(0.1).rgb().string(),
+              },
+              '& .MuiIconButton-root[aria-label="Show filters"]': {
+                color: `${COLORS.WHITE} !important`,
+              },
+            },
+            '& .MuiDataGrid-menuIcon': {
+              '& .MuiIconButton-root[aria-label="Menu"]': {
+                color: Color(COLORS.WHITE).fade(0.35).rgb().string(),
+              },
+            },
+            '&.MuiDataGrid-withBorder': {
+              borderRightColor: Color(COLORS.PRIMARY).lighten(0.2).hex(),
+            },
+          },
+          '& .MuiDataGrid-columnHeader--sorted': {
+            '& .MuiIconButton-root[aria-label="Sort"]': {
+              color: `${COLORS.WHITE} !important`,
+            },
+            backgroundColor: Color(COLORS.PRIMARY).darken(0.15).hex(),
+          },
+          '& .MuiDataGrid-columnHeaderWrapper': {
+            backgroundColor: COLORS.PRIMARY,
+          },
+        },
+      },
+      '&.secondary': {
+        '& .MuiDataGrid-main': {
+          '& .MuiDataGrid-columnHeader': {
+            '& .MuiDataGrid-columnHeaderTitleContainer': {
+              '& .MuiCheckbox-root': {
+                color: COLORS.WHITE,
+              },
+              '& .MuiDataGrid-columnHeaderTitle': {
+                color: COLORS.WHITE,
+              },
+            },
+            '& .MuiDataGrid-iconButtonContainer': {
+              '& .MuiIconButton-root': {
+                color: Color(COLORS.WHITE).fade(0.1).rgb().string(),
+              },
+              '& .MuiIconButton-root[aria-label="Show filters"]': {
+                color: `${COLORS.WHITE} !important`,
+              },
+            },
+            '& .MuiDataGrid-menuIcon': {
+              '& .MuiIconButton-root[aria-label="Menu"]': {
+                color: Color(COLORS.WHITE).fade(0.35).rgb().string(),
+              },
+            },
+            '&.MuiDataGrid-withBorder': {
+              borderRightColor: Color(COLORS.SECONDARY).lighten(0.5).hex(),
+            },
+          },
+          '& .MuiDataGrid-columnHeader--sorted': {
+            '& .MuiIconButton-root[aria-label="Sort"]': {
+              color: `${COLORS.WHITE} !important`,
+            },
+            backgroundColor: Color(COLORS.SECONDARY).darken(0.15).hex(),
+          },
+          '& .MuiDataGrid-columnHeaderWrapper': {
+            backgroundColor: COLORS.SECONDARY,
+          },
+        },
+      },
+      borderRadius: SHAPE.borderRadius,
+      overflow: 'hidden',
     },
   },
   MuiIconButton: {
@@ -127,7 +279,9 @@ const OVERRIDES: Overrides = {
         border: `6px solid ${COLORS.WHITE}`,
         color: COLORS.PRIMARY,
       },
+      height: 'unset',
       padding: 3,
+      width: 'unset',
     },
     thumb: {
       boxShadow: '0 1px 2px 1px rgba(0, 0, 0, 0.08)',
@@ -142,6 +296,6 @@ const OVERRIDES: Overrides = {
       opacity: 1,
     },
   },
-};
+} as Overrides;
 
 export default OVERRIDES;